+++ emr.el
@@ -73,19 +73,29 @@ If the defun is preceded by comments, move above them."
   (ignore-errors
     (beginning-of-thing 'defun))
   ;; If there is a comment attached to this defun, skip over it.
-  (while (save-excursion
-           (forward-line -1)
-           (and (emr-looking-at-comment?)
-                (not (bobp))))
-    (forward-line -1)))
+  (let* ((prev-line-pos
+          (unless (bobp)
+            (save-excursion
+              (forward-line -1)
+              (point))))
+         (comment-preceding
+          (and
+           prev-line-pos
+           (emr-looking-at-comment?
+            (line-end-position prev-line-pos)))))
+    (when comment-preceding
+      (forward-line -1)
+      (while (and (emr-looking-at-comment? (line-end-position))
+                  (not (bobp)))
+        (forward-line -1)))))
 
 (defun emr-looking-at-string? ()
   "Return non-nil if point is inside a string."
   (nth 3 (syntax-ppss)))
 
-(defun emr-looking-at-comment? ()
+(defun emr-looking-at-comment? (&optional pos)
   "Non-nil if point is on a comment."
-  (nth 4 (syntax-ppss)))
+  (nth 4 (syntax-ppss pos)))
 
 (defun emr-line-str ()
   "Return the contents of the current line."